Refactor project desktop components
