+++ src/features/project-desktop/components/project-home-overview.tsx
@@ -1,28 +1,26 @@
 import { useEffect, useMemo, useState } from 'react';
 import {
-  Check,
-  Copy,
-  ExternalLink,
-  Github,
   Grid2X2,
-  Info,
   List,
   Plus,
   RefreshCw,
-  Terminal,
-  X,
 } from 'lucide-react';
-import { Button, Chip, Tab, TabIndicator, TabList, Tabs, Text, Tooltip } from '@/app/dotnaos-ui';
+import {
+  Button,
+  Chip,
+  Tab,
+  TabIndicator,
+  TabList,
+  Tabs,
+  Text,
+  Tooltip
+} from '@/app/dotnaos-ui';
 import { projectSpaceClient } from '@/api/project-space-client';
 import type {
   ConnectorOverviewResult,
-  FullstackTemplateStatus,
-  GitHubCatalogRepository,
   GitHubCatalogResult,
   GitHubOAuthDeviceStartResult,
-  GitHubProjectConfigStatus,
   MachineRecord,
-  ProjectWorktreeRecord,
   ProjectSpaceRecord
 } from '@/shared/project-space-api';
 import {
@@ -32,6 +30,29 @@ import {
   MachineDeviceIcon,
   MachineOsMark
 } from './machine-visuals';
+import {
+  AddMachineDialog,
+  BranchChips,
+  GitHubConnectPanel,
+  MachineDetailsTooltip,
+  MainListSearch,
+  ProjectListItem
+} from './project-home-overview-widgets';
+import {
+  branchesFromWorktrees,
+  configChipClass,
+  getMachineId,
+  getProjectMachineId,
+  getTemplateStatus,
+  installScriptUrl,
+  isVisibleLocalProject,
+  matchesQuery,
+  mergeBranchChips,
+  projectMatchesRepo,
+  templateStatusLabels,
+  type BranchChipRecord,
+  type MatrixRow
+} from './project-home-overview-model';
 
 interface ProjectHomeOverviewProps {
   connector: ConnectorOverviewResult;
@@ -44,287 +65,6 @@ interface ProjectHomeOverviewProps {
   onSelectMachine(machineId: string): void;
   projects: ProjectSpaceRecord[];
   onSelectProject(projectId: string): void;
-}
-
-interface LocalMatch {
-  machineId: string;
-  project: ProjectSpaceRecord;
-}
-
-interface MatrixRow {
-  id: string;
-  isLocalOnly: boolean;
-  localMatches: LocalMatch[];
-  repo?: GitHubCatalogRepository;
-  title: string;
-}
-
-interface BranchChipRecord {
-  isBase: boolean;
-  name: string;
-}
-
-const templateStatusLabels: Record<FullstackTemplateStatus, string> = {
-  implemented: 'ok',
-  partial: 'partial',
-  'not-detected': 'missing',
-  'template-source': 'template'
-};
-
-function getMachineId(project: ProjectSpaceRecord) {
-  return project.id.includes(':') ? project.id.slice(0, project.id.indexOf(':')) : 'local';
-}
-
-function getProjectMachineId(project: ProjectSpaceRecord, localMachineId: string) {
-  const machineId = getMachineId(project);
-  return machineId === 'local' ? localMachineId : machineId;
-}
-
-function getTemplateStatus(project: ProjectSpaceRecord): FullstackTemplateStatus {
-  return project.fullstackTemplate?.status ?? 'not-detected';
-}
-
-function basename(path: string) {
-  return path.split('/').filter(Boolean).at(-1) ?? path;
-}
-
-function isVisibleLocalProject(project: ProjectSpaceRecord) {
-  if (project.kind === 'github') {
-    return false;
-  }
-
-  const projectFolder = basename(project.rootPath);
-
-  return !projectFolder.startsWith('.') && !projectFolder.endsWith('.worktrees');
-}
-
-function normalizeKey(value: string) {
-  return value.trim().replace(/^@/, '').toLowerCase();
-}
-
-function projectKeys(project: ProjectSpaceRecord) {
-  const name = normalizeKey(project.name);
-  const rootName = normalizeKey(basename(project.rootPath));
-  const keys = new Set([name, rootName]);
-
-  if (name.includes('/')) {
-    keys.add(name.split('/').at(-1) ?? name);
-  }
-
-  return keys;
-}
-
-function projectMatchesRepo(project: ProjectSpaceRecord, repo: GitHubCatalogRepository) {
-  const keys = projectKeys(project);
-  const fullName = normalizeKey(repo.fullName);
-  const name = normalizeKey(repo.name);
-
-  return keys.has(fullName) || keys.has(name);
-}
-
-function isBaseBranchName(branchName: string, defaultBranch?: string) {
-  const normalizedBranch = normalizeKey(branchName);
-  const normalizedDefault = defaultBranch ? normalizeKey(defaultBranch) : '';
-
-  return (
-    (normalizedDefault && normalizedBranch === normalizedDefault) ||
-    normalizedBranch === 'main' ||
-    normalizedBranch === 'master'
-  );
-}
-
-function branchesFromWorktrees(
-  worktrees: ProjectWorktreeRecord[],
-  defaultBranch?: string
-): BranchChipRecord[] {
-  const branches = new Map<string, BranchChipRecord>();
-
-  for (const worktree of worktrees) {
-    const branchName = worktree.branchName?.trim();
-
-    if (!branchName) {
-      continue;
-    }
-
-    branches.set(branchName, {
-      isBase: isBaseBranchName(branchName, defaultBranch),
-      name: branchName
-    });
-  }
-
-  return Array.from(branches.values()).sort((left, right) => {
-    if (left.isBase !== right.isBase) {
-      return left.isBase ? -1 : 1;
-    }
-
-    return left.name.localeCompare(right.name);
-  });
-}
-
-function mergeBranchChips(
-  defaultBranch: string | undefined,
-  branchGroups: BranchChipRecord[][]
-) {
-  const branches = new Map<string, BranchChipRecord>();
-
-  if (defaultBranch) {
-    branches.set(defaultBranch, {
-      isBase: true,
-      name: defaultBranch
-    });
-  }
-
-  for (const group of branchGroups) {
-    for (const branch of group) {
-      const existing = branches.get(branch.name);
-
-      branches.set(branch.name, {
-        isBase: existing?.isBase || branch.isBase || isBaseBranchName(branch.name, defaultBranch),
-        name: branch.name
-      });
-    }
-  }
-
-  return Array.from(branches.values()).sort((left, right) => {
-    if (left.isBase !== right.isBase) {
-      return left.isBase ? -1 : 1;
-    }
-
-    return left.name.localeCompare(right.name);
-  });
-}
-
-function configChipClass(status: GitHubProjectConfigStatus | FullstackTemplateStatus) {
-  if (status === 'complete' || status === 'implemented' || status === 'template-source') {
-    return 'text-emerald-300';
-  }
-
-  if (status === 'partial') {
-    return 'text-amber-300';
-  }
-
-  return 'text-neutral-500';
-}
-
-function formatLastSeen(value?: string) {
-  if (!value) {
-    return 'not seen';
-  }
-
-  const timestamp = Date.parse(value);
-
-  if (!Number.isFinite(timestamp)) {
-    return value;
-  }
-
-  const minutes = Math.max(0, Math.round((Date.now() - timestamp) / 60_000));
-
-  if (minutes < 1) {
-    return 'now';
-  }
-
-  if (minutes < 60) {
-    return `${minutes}m`;
-  }
-
-  const hours = Math.round(minutes / 60);
-  return hours < 24 ? `${hours}h` : `${Math.round(hours / 24)}d`;
-}
-
-function DetailRow({ label, mono, value }: { label: string; mono?: boolean; value: string }) {
-  return (
-    <div className="flex items-baseline justify-between gap-3">
-      <span className="shrink-0 text-[10px] font-medium text-neutral-500">{label}</span>
-      <span
-        className={[
-          'min-w-0 truncate text-right text-neutral-200',
-          mono ? 'font-mono text-[11px]' : 'text-xs'
-        ].join(' ')}
-      >
-        {value}
-      </span>
-    </div>
-  );
-}
-
-function installScriptUrl() {
-  if (typeof window === 'undefined') {
-    return '/connector/install.sh';
-  }
-
-  return `${window.location.origin}/connector/install.sh`;
-}
-
-function MachineDetailsTooltip({
-  machine,
-  projectCount
-}: {
-  machine: MachineRecord;
-  projectCount: number;
-}) {
-  const origin =
-    machine.connector.origin ?? machine.network.tailscaleIp ?? machine.connector.installCommand;
-
-  return (
-    <Tooltip delay={150}>
-      <Tooltip.Trigger className="inline-flex">
-        <span
-          aria-label={`${machine.name} details`}
-          className="inline-flex size-5 items-center justify-center rounded-full text-neutral-600 transition hover:text-neutral-300"
-        >
-          <Info className="size-3.5" strokeWidth={1.8} />
-        </span>
-      </Tooltip.Trigger>
-      <Tooltip.Content className="w-56 space-y-1.5">
-        {machine.connector.serviceName ? (
-          <DetailRow label="Service" value={machine.connector.serviceName} />
-        ) : null}
-        <DetailRow label="Status" value={machine.connector.status} />
-        <DetailRow label="Projects" value={String(projectCount)} />
-        <DetailRow label="Last seen" value={formatLastSeen(machine.connector.lastSeen)} />
-        {origin ? <DetailRow label="Origin" mono value={origin} /> : null}
-      </Tooltip.Content>
-    </Tooltip>
-  );
-}
-
-function BranchChips({ branches }: { branches: BranchChipRecord[] }) {
-  if (branches.length === 0) {
-    return null;
-  }
-
-  const visibleBranches = branches.slice(0, 3);
-  const hiddenCount = branches.length - visibleBranches.length;
-
-  return (
-    <div
-      aria-label="Branches"
-      className="flex max-w-[17rem] shrink-0 items-center gap-1 overflow-x-auto"
-    >
-      {visibleBranches.map((branch) => (
-        <span
-          key={branch.name}
-          title={branch.name}
-          className={[
-            'inline-flex max-w-28 shrink-0 items-center truncate rounded-full px-2 py-0.5 text-[11px] font-medium',
-            branch.isBase
-              ? 'bg-neutral-100 text-neutral-950'
-              : 'bg-neutral-800/80 text-neutral-300'
-          ].join(' ')}
-        >
-          {branch.name}
-        </span>
-      ))}
-      {hiddenCount > 0 ? (
-        <span
-          title={branches.slice(3).map((branch) => branch.name).join(', ')}
-          className="inline-flex shrink-0 rounded-full bg-neutral-900 px-2 py-0.5 text-[11px] font-medium text-neutral-500"
-        >
-          +{hiddenCount}
-        </span>
-      ) : null}
-    </div>
-  );
 }
 
 export function ProjectHomeOverview({
@@ -350,6 +90,8 @@ export function ProjectHomeOverview({
   const [installScriptHref, setInstallScriptHref] = useState('/connector/install.sh');
   const [hasCopiedInstallCommand, setHasCopiedInstallCommand] = useState(false);
   const [layout, setLayout] = useState<'grid' | 'list'>('list');
+  const [machineQuery, setMachineQuery] = useState('');
+  const [projectQuery, setProjectQuery] = useState('');
   const [selectedMachineId, setSelectedMachineId] = useState('');
 
   async function refresh(includeConnector: boolean) {
@@ -438,8 +180,29 @@ export function ProjectHomeOverview({
     machines.find((machine) => machine.connector.status === 'local')?.id ??
     machines[0]?.id ??
     'local';
-  const connectedMachines = machines.filter(isMachineConnected);
-  const disconnectedMachines = machines.filter((machine) => !isMachineConnected(machine));
+  const filteredMachines = useMemo(
+    () =>
+      machines.filter((machine) =>
+        matchesQuery(
+          [
+            machine.name,
+            machine.id,
+            machine.kind,
+            machine.profile,
+            machine.primaryUser,
+            machine.network.localName,
+            machine.network.sshUser,
+            machine.network.tailscaleIp,
+            machine.connector.serviceName,
+            machine.connector.status
+          ],
+          machineQuery
+        )
+      ),
+    [machineQuery, machines]
+  );
+  const connectedMachines = filteredMachines.filter(isMachineConnected);
+  const disconnectedMachines = filteredMachines.filter((machine) => !isMachineConnected(machine));
   const activeMachineId = selectedMachineId || localMachineId || machines[0]?.id || '';
   const activeMachine = machinesById.get(activeMachineId);
 
@@ -499,6 +262,23 @@ export function ProjectHomeOverview({
   }, [localMachineId, projects]);
   const activeMachineProjects = activeMachineId ? projectsByMachine[activeMachineId] ?? [] : [];
   const githubRows = rows.filter((row) => !row.isLocalOnly);
+  const filteredGithubRows = useMemo(
+    () =>
+      githubRows.filter((row) =>
+        matchesQuery(
+          [
+            row.title,
+            row.repo?.fullName,
+            row.repo?.name,
+            row.repo?.owner,
+            row.repo?.description,
+            row.localMatches.map((match) => match.project.name).join(' ')
+          ],
+          projectQuery
+        )
+      ),
+    [githubRows, projectQuery]
+  );
   const branchSourceProjects = useMemo(() => {
     if (mode !== 'projects') {
       return [];
@@ -519,7 +299,7 @@ export function ProjectHomeOverview({
   const githubRowGroups = useMemo(() => {
     const groups = new Map<string, MatrixRow[]>();
 
-    for (const row of githubRows) {
+    for (const row of filteredGithubRows) {
       const owner = row.repo?.owner ?? 'Other';
       const items = groups.get(owner) ?? [];
       items.push(row);
@@ -532,7 +312,7 @@ export function ProjectHomeOverview({
         owner
       }))
       .sort((left, right) => left.owner.localeCompare(right.owner));
-  }, [githubRows]);
+  }, [filteredGithubRows]);
 
   useEffect(() => {
     const missingProjects = branchSourceProjects.filter(
@@ -737,197 +517,17 @@ export function ProjectHomeOverview({
     );
   }
 
-  function projectIdForRow(row: MatrixRow) {
-    return row.localMatches[0]?.project.id ?? (row.repo ? `github:${row.repo.fullName}` : '');
-  }
-
-  function renderGitHubLink(repo: GitHubCatalogRepository, label: string) {
-    return (
-      <a
-        href={repo.url}
-        target="_blank"
-        rel="noreferrer"
-        aria-label={`Open ${repo.fullName} on GitHub`}
-        onClick={(event) => event.stopPropagation()}
-        className="inline-flex size-8 shrink-0 items-center justify-center rounded-lg text-neutral-600 transition hover:bg-neutral-900 hover:text-neutral-200"
-        title={label}
-      >
-        <ExternalLink className="size-3.5" />
-      </a>
-    );
-  }
-
-  function renderProjectRow(row: MatrixRow) {
-    const branches = branchesForRow(row);
-    const projectId = projectIdForRow(row);
-
-    return (
-      <div
-        key={row.id}
-        className="flex min-w-0 items-center gap-3 rounded-lg px-3 py-2 transition hover:bg-neutral-900/40"
-      >
-        <button
-          type="button"
-          onClick={() => projectId && onSelectProject(projectId)}
-          className="min-w-0 flex-1 text-left"
-        >
-          <Text className="block min-w-0 truncate text-sm font-medium text-neutral-100">
-            {row.repo?.name ?? row.title}
-          </Text>
-        </button>
-        <BranchChips branches={branches} />
-        {row.repo ? renderGitHubLink(row.repo, 'Open on GitHub') : null}
-      </div>
-    );
-  }
-
-  function renderProjectCard(row: MatrixRow) {
-    const branches = branchesForRow(row);
-    const projectId = projectIdForRow(row);
-
-    return (
-      <div
-        key={row.id}
-        className="group flex min-w-0 flex-col gap-4 rounded-lg border border-neutral-900 bg-neutral-950/45 p-4 transition hover:border-neutral-800 hover:bg-neutral-950/70"
-      >
-        <div className="flex min-w-0 items-start justify-between gap-3">
-          <button
-            type="button"
-            onClick={() => projectId && onSelectProject(projectId)}
-            className="min-w-0 flex-1 text-left"
-          >
-            <Text className="block min-w-0 truncate text-sm font-semibold text-neutral-100">
-              {row.repo?.name ?? row.title}
-            </Text>
-            <Text className="mt-1 block truncate text-xs text-neutral-500">
-              {row.repo?.owner ?? 'Local'}
-            </Text>
-          </button>
-          {row.repo ? renderGitHubLink(row.repo, 'Open on GitHub') : null}
-        </div>
-        <BranchChips branches={branches} />
-      </div>
-    );
-  }
-
-  function renderGitHubConnectPanel() {
-    if (githubCatalog.status === 'connected') {
-      return null;
-    }
-
-    return (
-      <div className="mb-4 rounded-lg bg-neutral-950/60 px-4 py-4">
-        <div className="flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between">
-          <div className="min-w-0">
-            <Text className="block text-sm font-semibold text-neutral-100">
-              Connect GitHub
-            </Text>
-            <Text className="mt-1 block text-sm text-neutral-500">
-              {githubFlow?.status === 'pending'
-                ? 'Enter this code on GitHub, then check the login here.'
-                : githubCatalog.message ?? 'Connect GitHub to load repositories.'}
-            </Text>
-          </div>
-          {githubFlow?.status === 'pending' ? (
-            <div className="flex shrink-0 flex-wrap items-center gap-2">
-              <span className="rounded-lg bg-neutral-900 px-3 py-1.5 font-mono text-sm font-semibold text-neutral-100">
-                {githubFlow.userCode}
-              </span>
-              {githubFlow.verificationUri ? (
-                <a href={githubFlow.verificationUri} target="_blank" rel="noreferrer">
-                  <Button size="sm" variant="outline">
-                    <ExternalLink className="size-4" />
-                    Open GitHub
-                  </Button>
-                </a>
-              ) : null}
-              <Button size="sm" isDisabled={isConnectingGitHub} onPress={() => void pollGitHubLogin()}>
-                Check login
-              </Button>
-            </div>
-          ) : (
-            <Button
-              className="shrink-0"
-              size="sm"
-              isDisabled={isConnectingGitHub || githubCatalog.status === 'not-configured'}
-              onPress={() => void connectGitHub()}
-            >
-              <Github className="size-4" />
-              Login with GitHub
-            </Button>
-          )}
-        </div>
-      </div>
-    );
-  }
-
-  function renderInstallDialog() {
-    if (!isInstallDialogOpen) {
-      return null;
-    }
-
-    return (
-      <div className="fixed inset-0 z-50 flex items-end bg-black/70 p-4 sm:items-center sm:justify-center">
-        <div className="w-full max-w-xl rounded-lg border border-neutral-800 bg-neutral-950 p-4 shadow-2xl">
-          <div className="flex items-start justify-between gap-4">
-            <div className="min-w-0">
-              <Text className="block text-base font-semibold text-neutral-100">
-                Add a machine
-              </Text>
-              <Text className="mt-1 block text-sm text-neutral-500">
-                Run this on the Mac you want to add. It installs the connector and keeps it running.
-              </Text>
-            </div>
-            <Button
-              aria-label="Close add machine"
-              isIconOnly
-              size="sm"
-              variant="ghost"
-              onPress={() => setIsInstallDialogOpen(false)}
-            >
-              <X className="size-4" />
-            </Button>
-          </div>
-
-          <div className="mt-4 rounded-lg bg-black px-3 py-3">
-            <div className="mb-2 flex items-center gap-2 text-xs font-medium text-neutral-500">
-              <Terminal className="size-3.5" />
-              macOS arm64
-            </div>
-            <code className="block whitespace-pre-wrap break-all font-mono text-sm text-neutral-100">
-              {installCommand}
-            </code>
-          </div>
-
-          <div className="mt-4 flex flex-wrap gap-2">
-            <Button size="sm" onPress={() => void copyInstallCommand()}>
-              {hasCopiedInstallCommand ? <Check className="size-4" /> : <Copy className="size-4" />}
-              {hasCopiedInstallCommand ? 'Copied' : 'Copy command'}
-            </Button>
-            <a href={installScriptHref} target="_blank" rel="noreferrer">
-              <Button size="sm" variant="outline">
-                <ExternalLink className="size-4" />
-                Open script
-              </Button>
-            </a>
-            <a href="/connector" target="_blank" rel="noreferrer">
-              <Button size="sm" variant="ghost">
-                Install guide
-              </Button>
-            </a>
-          </div>
-
-          <Text className="mt-3 block text-xs text-neutral-600">
-            Linux packaging is not published yet. Build from source or run a matching connector binary with PROJECT_CONNECTOR_HUB_URL set to this site.
-          </Text>
-        </div>
-      </div>
-    );
-  }
-
   return (
     <div className="flex min-h-full w-full flex-col">
-      {renderInstallDialog()}
+      {isInstallDialogOpen ? (
+        <AddMachineDialog
+          hasCopiedInstallCommand={hasCopiedInstallCommand}
+          installCommand={installCommand}
+          installScriptHref={installScriptHref}
+          onClose={() => setIsInstallDialogOpen(false)}
+          onCopy={() => void copyInstallCommand()}
+        />
+      ) : null}
       <div className="mb-3 flex items-center justify-between gap-3">
         <div className="min-w-0">
           {mode === 'machines' ? (
@@ -968,11 +568,36 @@ export function ProjectHomeOverview({
         </div>
       </div>
 
+      {mode === 'machines' || githubCatalog.status === 'connected' ? (
+        <div className="mb-4">
+          {mode === 'machines' ? (
+            <MainListSearch
+              label="Search machines"
+              placeholder="Search machines"
+              value={machineQuery}
+              onChange={setMachineQuery}
+            />
+          ) : (
+            <MainListSearch
+              label="Search projects"
+              placeholder="Search projects"
+              value={projectQuery}
+              onChange={setProjectQuery}
+            />
+          )}
+        </div>
+      ) : null}
+
       {mode === 'machines' ? (
         <>
           <div className="mb-4 space-y-4">
             {renderMachineSection('Connected', connectedMachines)}
             {renderMachineSection('Disconnected', disconnectedMachines)}
+            {filteredMachines.length === 0 ? (
+              <div className="rounded-lg bg-neutral-950/45 px-4 py-6">
+                <Text className="text-sm text-neutral-500">No machines found.</Text>
+              </div>
+            ) : null}
           </div>
 
           <div className="mb-3">
@@ -1019,7 +644,15 @@ export function ProjectHomeOverview({
         </>
       ) : null}
 
-      {mode === 'projects' ? renderGitHubConnectPanel() : null}
+      {mode === 'projects' ? (
+        <GitHubConnectPanel
+          flow={githubFlow}
+          githubCatalog={githubCatalog}
+          isConnecting={isConnectingGitHub}
+          onConnect={() => void connectGitHub()}
+          onPoll={() => void pollGitHubLogin()}
+        />
+      ) : null}
 
       {mode === 'projects' ? (
         <div className="flex max-h-[70vh] flex-col overflow-y-auto">
@@ -1029,10 +662,23 @@ export function ProjectHomeOverview({
                 {group.owner}
               </Text>
               <div className={layout === 'grid' ? 'grid gap-3 md:grid-cols-2 xl:grid-cols-3' : 'flex flex-col'}>
-                {group.items.map(layout === 'grid' ? renderProjectCard : renderProjectRow)}
+                {group.items.map((row) => (
+                  <ProjectListItem
+                    key={row.id}
+                    branches={branchesForRow(row)}
+                    layout={layout}
+                    onSelectProject={onSelectProject}
+                    row={row}
+                  />
+                ))}
               </div>
             </div>
           ))}
+          {githubCatalog.status === 'connected' && filteredGithubRows.length === 0 ? (
+            <div className="rounded-lg bg-neutral-950/45 px-4 py-6">
+              <Text className="text-sm text-neutral-500">No projects found.</Text>
+            </div>
+          ) : null}
         </div>
       ) : null}
     </div>
